+++ R/deparse.R
@@ -1030,8 +1030,10 @@ call_deparse_highlight <- function(call, arg) {
     args_list <- sprintf("%s = %s", arg, as_label(call[[arg]]))
     args_list <- format_arg_unquoted(args_list)
   } else {
-    args_list <- paste_line(args_deparse(node_cdr(call)))
-    args_list <- substring(args_list, 2, nchar(args_list) - 1)
+    args_list <- call
+    args_list[[1]] <- quote(F)
+    args_list <- as_label(args_list)
+    args_list <- substring(args_list, 3, nchar(args_list) - 1)
   }
 
   head <- call[[1]]