+++ R/rlang-package.R
@@ -3,16 +3,6 @@
 "_PACKAGE"
 
 on_load({
-  # TODO! Once pkgload is on CRAN
-  # check_downstream("1.0.3", "pkgload (>= 1.2.5)")
-  check_downstream("0.4.0", "dplyr (>= 0.8.0)")
-  check_downstream(
-    "1.0.0",
-    "ellipsis (>= 0.3.2)",
-    "vctrs (>= 0.3.8)",
-    info = "Not updating now is completely safe and will only cause import warnings."
-  )
-
   local_use_cli()
 })